+++ app/src/main/java/com/example/androidcomposebeginner/MainActivity.kt
@@ -6,9 +6,7 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -16,7 +14,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.example.androidcomposebeginner.ui.theme.AndroidComposeBeginnerTheme
 
-class RowActivity : ComponentActivity() {
+class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
@@ -26,26 +24,9 @@ class RowActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    RowComponent(modifier = Modifier.fillMaxSize())
+
                 }
             }
         }
-    }
-}
-
-@Composable
-fun RowComponent(modifier: Modifier = Modifier) {
-    Row (modifier = modifier){
-        Text(text = "No")
-        Text(text = "Name")
-        Text(text = "Age")
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun RowPreview() {
-    AndroidComposeBeginnerTheme {
-        RowComponent(modifier = Modifier.fillMaxSize())
     }
 }
